lib: Use BTreeMap to store mapping from integers, instead of HashMap
Fixes #18

Mapping _from_ integers should be performed in a deterministic,
incrementing order so that it is more clear when iterating the bindings
(ie. in dbg prints), and some drivers appear to choke (#18) on
descriptor layouts that are created with the objects in a different
order, even if the resulting layout is the same.